pyproject.toml
[tool.poetry]
name = "web-hw-14"
version = "0.1.0"
description = ""
authors = ["alexfighterbomber <ssamsobi@gmail.com>"]
readme = "README.md"
packages = [{include = "web_hw_14"}]

[tool.poetry.dependencies]
python = "^3.13"
greenlet = "3.2.3"
sqlalchemy = "^2.0.0"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
psycopg = {extras = ["binary", "pool"], version = "^3.2.9"}
alembic = "^1.9.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
libgravatar = "^1.0.3"
fastapi-mail = "^1.2.5"
redis = "^4.5.1"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.32.0"
bcrypt = "^4.0.1"
starlette = "^0.47.2"
pydantic = "<2.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"


[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
pytest = "^8.4.1"
aiosqlite = "^0.21.0"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
pytest-xdist = "^3.8.0"
asgi-lifespan = "^2.1.0"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
